Refactor market list modules: extract common logic to MarketList module and use it in MarketDiscovery and MarketWatchlist modules
