+++ UnstoppableWallet/UnstoppableWallet/Modules/Market/MarketDiscovery/MarketDiscoveryModule.swift
@@ -2,10 +2,13 @@ struct MarketDiscoveryModule {
 
     static func viewController(marketViewModel: MarketViewModel) -> MarketDiscoveryViewController {
         let categoriesProvider = MarketCategoriesProvider()
-        let service = MarketDiscoveryService(currencyKit: App.shared.currencyKit, rateManager: App.shared.rateManager, categoriesProvider: categoriesProvider)
+        let service = MarketDiscoveryService(rateManager: App.shared.rateManager, categoriesProvider: categoriesProvider)
+        let listService = MarketListService(currencyKit: App.shared.currencyKit, fetcher: service)
 
         let viewModel = MarketDiscoveryViewModel(service: service)
-        return MarketDiscoveryViewController(marketViewModel: marketViewModel, viewModel: viewModel)
+        let listViewModel = MarketListViewModel(service: listService)
+
+        return MarketDiscoveryViewController(marketViewModel: marketViewModel, listViewModel: listViewModel, viewModel: viewModel)
     }
 
 }
