+++ UnstoppableWallet/UnstoppableWallet/Modules/Market/MarketWatchlist/MarketWatchlistModule.swift
@@ -3,10 +3,11 @@ import CurrencyKit
 struct MarketWatchlistModule {
 
     static func viewController() -> MarketWatchlistViewController {
-        let service = MarketWatchlistService(currencyKit: App.shared.currencyKit, rateManager: App.shared.rateManager, favoritesManager: App.shared.favoritesManager)
+        let service = MarketWatchlistService(rateManager: App.shared.rateManager, favoritesManager: App.shared.favoritesManager)
+        let listService = MarketListService(currencyKit: App.shared.currencyKit, fetcher: service)
 
-        let viewModel = MarketWatchlistViewModel(service: service)
-        return MarketWatchlistViewController(viewModel: viewModel)
+        let listViewModel = MarketListViewModel(service: listService)
+        return MarketWatchlistViewController(listViewModel: listViewModel)
     }
 
 }
